scanner: move Matcher code to separate file
